+++ src/navigator.js
@@ -8,10 +8,11 @@ export default class Navigator {
         this.url = document.location.href;
         this.loaders = [];
         this.handler = handler;
+        this.events = { beforeFilter: [], beforeLoad: [], load: [], error: [] };
         this.filters = [
             (el, url) => url && url.indexOf(`${document.location.protocol}//${document.location.host}`) === 0,
             (el, url) => el.tagName === 'FORM' || !isAnchor(url),
-            (el) => !el.target,
+            (el, url, submitter) => !el.target && (!submitter || submitter.target),
             (el) => !el.hasAttribute('download'),
         ];
     }
@@ -31,6 +32,32 @@ export default class Navigator {
     }
 
     /**
+     * Add an event
+     *
+     * @param {string} event
+     * @param {Function} handler
+     *
+     * @return {this}
+     */
+    on(event, handler) {
+        this.events[event].push(handler);
+
+        return this;
+    }
+
+    /**
+     * Trigger an event
+     *
+     * @param {string} event
+     * @param  {...any} args
+     *
+     * @return {bool}
+     */
+    trigger(event, ...args) {
+        return this.events[event].every((handler) => handler(...args) !== false);
+    }
+
+    /**
      * Init the navigator, attach the events to capture the history changes
      *
      * @return {this}
@@ -42,47 +69,49 @@ export default class Navigator {
                 !event.ctrlKey &&
                 !event.altKey &&
                 !event.metaKey &&
+                this.trigger('beforeFilter', link, link.href) &&
                 this.filters.every((filter) => filter(link, link.href))
             ) {
-                this.go(link.href, event);
+                this.go(link.href, event, link);
                 event.preventDefault();
             }
         });
 
-        let latestBtn = null;
+        let getSubmitter;
 
-        delegate('click', 'form button,input[type="submit"]', (event, button) => (latestBtn = button));
+        if (window.SubmitEvent) {
+            getSubmitter = (event) => event.submitter;
+        } else {
+            let submitter;
+            delegate('click', 'form button,input[type="submit"]', (event, button) => (submitter = button));
+
+            getSubmitter = (event, form) => {
+                if (!submitter) {
+                    return null;
+                }
+
+                if (form.contains(submitter)) {
+                    const tmp = submitter;
+                    submitter = null;
+                    return tmp;
+                }
+
+                submitter = null;
+                return null;
+            };
+        }
 
         delegate('submit', 'form', (event, form) => {
-            if (latestBtn) {
-                if (!form.contains(latestBtn)) {
-                    latestBtn = null;
-                } else if (latestBtn.formTarget) {
-                    latestBtn = null;
-                    return;
-                }
-            }
+            const submitter = getSubmitter(event, form);
+            const url = getFormUrl(form, submitter);
 
-            const url = resolve(latestBtn && latestBtn.hasAttribute('formaction') ? latestBtn.formAction : form.action);
-
-            if (this.filters.every((filter) => filter(form, url))) {
-                const options = { url };
-
-                if (latestBtn) {
-                    if (latestBtn.name) {
-                        options.body = new FormData(form);
-                        options.body.append(latestBtn.name, latestBtn.value);
-                    }
-                    if (latestBtn.hasAttribute('formmethod')) {
-                        options.method = latestBtn.formMethod.toUpperCase();
-                    }
-                }
-
-                this.submit(form, event, options);
+            if (
+                this.trigger('beforeFilter', form, url, submitter) &&
+                this.filters.every((filter) => filter(form, url, submitter))
+            ) {
+                this.submit(form, event, { url }, submitter);
                 event.preventDefault();
             }
-
-            latestBtn = null;
         });
 
         window.addEventListener('popstate', (event) => {
@@ -102,10 +131,12 @@ export default class Navigator {
      * @param  {string} url
      * @param  {Event} event
      * @param  {Object} options
+     * @param  {HTMLElement} element
+     * @param  {HTMLElement} submitter
      *
      * @return {Promise}
      */
-    go(url, event, options = {}) {
+    go(url, event, options = {}, element, submitter) {
         url = resolve(url);
 
         let loader = this.loaders.find((loader) => loader.url === url);
@@ -115,7 +146,7 @@ export default class Navigator {
             this.loaders.push(loader);
         }
 
-        return this.load(loader, event);
+        return this.load(loader, event, element, submitter);
     }
 
     /**
@@ -124,11 +155,25 @@ export default class Navigator {
      * @param  {HTMLFormElement} form
      * @param  {Event} event
      * @param  {Object} options
+     * @param  {HTMLElement} submitter
      *
      * @return {Promise}
      */
-    submit(form, event, options = {}) {
-        return this.load(new FormLoader(form, options), event);
+    submit(form, event, options = {}, submitter) {
+        options.url = options.url || getFormUrl(form, submitter);
+
+        if (submitter) {
+            if (submitter.name) {
+                options.body = new FormData(form);
+                options.body.append(submitter.name, submitter.value);
+            }
+
+            if (submitter.hasAttribute('formmethod')) {
+                options.method = submitter.formMethod.toUpperCase();
+            }
+        }
+
+        return this.load(new FormLoader(form, options), event, form, submitter);
     }
 
     /**
@@ -136,22 +181,33 @@ export default class Navigator {
      *
      * @param  {UrlLoader|FormLoader} loader
      * @param  {Event} event
+     * @param  {HTMLElement} element
+     * @param  {HTMLElement} submitter
      *
      * @return {Promise}
      */
-    load(loader, event) {
+    load(loader, event, element, submitter) {
         this.url = loader.url;
 
         const onError = (err) => {
-            console.error(err);
+            if (!this.trigger('error', err, element, loader, event, submitter)) {
+                return;
+            }
+
             loader.fallback();
         };
+
+        if (!this.trigger('beforeLoad', element, loader, event, submitter)) {
+            return;
+        }
 
         try {
             const result = this.handler(() => loader.load(), event);
 
             if (result instanceof Promise) {
-                return result.catch(onError);
+                return result.catch(onError).then(() => this.trigger('beforeLoad', element, loader, event, submitter));
+            } else {
+                this.trigger('beforeLoad', element, loader, event, submitter);
             }
         } catch (err) {
             onError(err);
@@ -184,4 +240,8 @@ function delegate(event, selector, callback) {
 
 function isAnchor(url) {
     return url.split('#', 1).shift() === document.location.href.split('#', 1).shift();
+}
+
+function getFormUrl(form, submitter) {
+    return resolve(submitter && submitter.hasAttribute('formaction') ? submitter.formAction : form.action);
 }